HTCONDOR-2610 fixup time_t printf formats for systems where time_t is a long long...

+++ src/condor_daemon_client/dc_transfer_queue.cpp
@@ -425,8 +425,8 @@ DCTransferQueue::SendReport(time_t now,bool disconnect)
 	if( interval < 0 ) {
 		interval = 0;
 	}
-	formatstr(report,"%ld %u %u %u %u %u %u %u",
-			  now,
+	formatstr(report,"%lld %u %u %u %u %u %u %u",
+			  (long long)now,
 			  (unsigned)interval,
 			  m_recent_bytes_sent,
 			  m_recent_bytes_received,
